Revert "feat(start-api): CloudFormation AWS::ApiGateway::RestApi support (#1238)" (#1282)
This reverts commit 7cb50b0174f85684e0e0e8ab15e056a309b6504b.
