+++ tests/unit/commands/local/lib/test_sam_api_provider.py
@@ -7,9 +7,27 @@ from nose_parameterized import parameterized
 
 from six import assertCountEqual
 
-from samcli.commands.local.lib.api_provider import ApiProvider, SamApiProvider
+from samcli.commands.local.lib.sam_api_provider import SamApiProvider
 from samcli.commands.local.lib.provider import Api
 from samcli.commands.validate.lib.exceptions import InvalidSamDocumentException
+
+
+class TestSamApiProvider_init(TestCase):
+
+    @patch.object(SamApiProvider, "_extract_apis")
+    @patch("samcli.commands.local.lib.sam_api_provider.SamBaseProvider")
+    def test_provider_with_valid_template(self, SamBaseProviderMock, extract_api_mock):
+        extract_api_mock.return_value = {"set", "of", "values"}
+
+        template = {"Resources": {"a": "b"}}
+        SamBaseProviderMock.get_template.return_value = template
+
+        provider = SamApiProvider(template)
+
+        self.assertEquals(len(provider.apis), 3)
+        self.assertEquals(provider.apis, set(["set", "of", "values"]))
+        self.assertEquals(provider.template_dict, {"Resources": {"a": "b"}})
+        self.assertEquals(provider.resources, {"a": "b"})
 
 
 class TestSamApiProviderWithImplicitApis(TestCase):
@@ -24,8 +42,9 @@ class TestSamApiProviderWithImplicitApis(TestCase):
             }
         }
 
-        provider = ApiProvider(template)
+        provider = SamApiProvider(template)
 
+        self.assertEquals(len(provider.apis), 0)
         self.assertEquals(provider.apis, [])
 
     @parameterized.expand([("GET"), ("get")])
@@ -53,7 +72,7 @@ class TestSamApiProviderWithImplicitApis(TestCase):
             }
         }
 
-        provider = ApiProvider(template)
+        provider = SamApiProvider(template)
 
         self.assertEquals(len(provider.apis), 1)
         self.assertEquals(list(provider.apis)[0], Api(path="/path", method="GET", function_name="SamFunc1", cors=None,
@@ -90,7 +109,7 @@ class TestSamApiProviderWithImplicitApis(TestCase):
             }
         }
 
-        provider = ApiProvider(template)
+        provider = SamApiProvider(template)
 
         api_event1 = Api(path="/path", method="GET", function_name="SamFunc1", cors=None, stage_name="Prod")
         api_event2 = Api(path="/path", method="POST", function_name="SamFunc1", cors=None, stage_name="Prod")
@@ -140,7 +159,7 @@ class TestSamApiProviderWithImplicitApis(TestCase):
             }
         }
 
-        provider = ApiProvider(template)
+        provider = SamApiProvider(template)
 
         api1 = Api(path="/path", method="GET", function_name="SamFunc1", cors=None, stage_name="Prod")
         api2 = Api(path="/path", method="POST", function_name="SamFunc2", cors=None, stage_name="Prod")
@@ -171,7 +190,7 @@ class TestSamApiProviderWithImplicitApis(TestCase):
             }
         }
 
-        provider = ApiProvider(template)
+        provider = SamApiProvider(template)
 
         self.assertEquals(provider.apis, [])
 
@@ -190,7 +209,7 @@ class TestSamApiProviderWithImplicitApis(TestCase):
             }
         }
 
-        provider = ApiProvider(template)
+        provider = SamApiProvider(template)
 
         self.assertEquals(provider.apis, [])
 
@@ -235,7 +254,7 @@ class TestSamApiProviderWithImplicitApis(TestCase):
             }
         }
 
-        provider = ApiProvider(template)
+        provider = SamApiProvider(template)
 
         result = [f for f in provider.get_all()]
 
@@ -248,7 +267,7 @@ class TestSamApiProviderWithImplicitApis(TestCase):
     def test_provider_get_all_with_no_apis(self):
         template = {}
 
-        provider = ApiProvider(template)
+        provider = SamApiProvider(template)
 
         result = [f for f in provider.get_all()]
 
@@ -279,7 +298,7 @@ class TestSamApiProviderWithImplicitApis(TestCase):
             }
         }
 
-        provider = ApiProvider(template)
+        provider = SamApiProvider(template)
 
         api_get = Api(path="/path", method="GET", function_name="SamFunc1", cors=None, stage_name="Prod")
         api_post = Api(path="/path", method="POST", function_name="SamFunc1", cors=None, stage_name="Prod")
@@ -332,7 +351,7 @@ class TestSamApiProviderWithImplicitApis(TestCase):
             }
         }
 
-        provider = ApiProvider(template)
+        provider = SamApiProvider(template)
 
         self.assertEquals(len(provider.apis), 1)
         self.assertEquals(list(provider.apis)[0], Api(path="/path", method="GET", function_name="SamFunc1",
@@ -384,7 +403,7 @@ class TestSamApiProviderWithImplicitApis(TestCase):
             Api(path="/path", method="PATCH", function_name="SamFunc1", binary_media_types=binary, stage_name="Prod")
         ]
 
-        provider = ApiProvider(template)
+        provider = SamApiProvider(template)
 
         assertCountEqual(self, provider.apis, expected_apis)
 
@@ -429,8 +448,9 @@ class TestSamApiProviderWithExplicitApis(TestCase):
             }
         }
 
-        provider = ApiProvider(template)
+        provider = SamApiProvider(template)
 
+        self.assertEquals(len(provider.apis), 0)
         self.assertEquals(provider.apis, [])
 
     def test_with_inline_swagger_apis(self):
@@ -447,7 +467,7 @@ class TestSamApiProviderWithExplicitApis(TestCase):
             }
         }
 
-        provider = ApiProvider(template)
+        provider = SamApiProvider(template)
         assertCountEqual(self, self.input_apis, provider.apis)
 
     def test_with_swagger_as_local_file(self):
@@ -471,11 +491,11 @@ class TestSamApiProviderWithExplicitApis(TestCase):
                 }
             }
 
-            provider = ApiProvider(template)
+            provider = SamApiProvider(template)
             assertCountEqual(self, self.input_apis, provider.apis)
 
-    @patch("samcli.commands.local.lib.cfn_base_api_provider.SamSwaggerReader")
-    def test_with_swagger_as_both_body_and_uri_called(self, SamSwaggerReaderMock):
+    @patch("samcli.commands.local.lib.sam_api_provider.SamSwaggerReader")
+    def test_with_swagger_as_both_body_and_uri(self, SamSwaggerReaderMock):
         body = {"some": "body"}
         filename = "somefile.txt"
 
@@ -496,7 +516,7 @@ class TestSamApiProviderWithExplicitApis(TestCase):
         SamSwaggerReaderMock.return_value.read.return_value = make_swagger(self.input_apis)
 
         cwd = "foo"
-        provider = ApiProvider(template, cwd=cwd)
+        provider = SamApiProvider(template, cwd=cwd)
         assertCountEqual(self, self.input_apis, provider.apis)
         SamSwaggerReaderMock.assert_called_with(definition_body=body, definition_uri=filename, working_dir=cwd)
 
@@ -527,7 +547,7 @@ class TestSamApiProviderWithExplicitApis(TestCase):
             }
         }
 
-        provider = ApiProvider(template)
+        provider = SamApiProvider(template)
         assertCountEqual(self, expected_apis, provider.apis)
 
     def test_with_binary_media_types(self):
@@ -560,7 +580,7 @@ class TestSamApiProviderWithExplicitApis(TestCase):
                 binary_media_types=expected_binary_types, stage_name="Prod")
         ]
 
-        provider = ApiProvider(template)
+        provider = SamApiProvider(template)
         assertCountEqual(self, expected_apis, provider.apis)
 
     def test_with_binary_media_types_in_swagger_and_on_resource(self):
@@ -589,7 +609,7 @@ class TestSamApiProviderWithExplicitApis(TestCase):
                 stage_name="Prod"),
         ]
 
-        provider = ApiProvider(template)
+        provider = SamApiProvider(template)
         assertCountEqual(self, expected_apis, provider.apis)
 
 
@@ -666,7 +686,7 @@ class TestSamApiProviderWithExplicitAndImplicitApis(TestCase):
             Api(path="/path3", method="POST", function_name="ImplicitFunc", cors=None, stage_name="Prod")
         ]
 
-        provider = ApiProvider(self.template)
+        provider = SamApiProvider(self.template)
         assertCountEqual(self, expected_apis, provider.apis)
 
     def test_must_prefer_implicit_api_over_explicit(self):
@@ -703,7 +723,7 @@ class TestSamApiProviderWithExplicitAndImplicitApis(TestCase):
             Api(path="/path3", method="GET", function_name="explicitfunction", cors=None, stage_name="Prod"),
         ]
 
-        provider = ApiProvider(self.template)
+        provider = SamApiProvider(self.template)
         assertCountEqual(self, expected_apis, provider.apis)
 
     def test_must_prefer_implicit_with_any_method(self):
@@ -737,7 +757,7 @@ class TestSamApiProviderWithExplicitAndImplicitApis(TestCase):
             Api(path="/path", method="PATCH", function_name="ImplicitFunc", cors=None, stage_name="Prod")
         ]
 
-        provider = ApiProvider(self.template)
+        provider = SamApiProvider(self.template)
         assertCountEqual(self, expected_apis, provider.apis)
 
     def test_with_any_method_on_both(self):
@@ -782,7 +802,8 @@ class TestSamApiProviderWithExplicitAndImplicitApis(TestCase):
             Api(path="/path2", method="POST", function_name="explicitfunction", cors=None, stage_name="Prod")
         ]
 
-        provider = ApiProvider(self.template)
+        provider = SamApiProvider(self.template)
+        print(provider.apis)
         assertCountEqual(self, expected_apis, provider.apis)
 
     def test_must_add_explicit_api_when_ref_with_rest_api_id(self):
@@ -819,7 +840,7 @@ class TestSamApiProviderWithExplicitAndImplicitApis(TestCase):
             Api(path="/newpath2", method="POST", function_name="ImplicitFunc", cors=None, stage_name="Prod")
         ]
 
-        provider = ApiProvider(self.template)
+        provider = SamApiProvider(self.template)
         assertCountEqual(self, expected_apis, provider.apis)
 
     def test_both_apis_must_get_binary_media_types(self):
@@ -874,7 +895,7 @@ class TestSamApiProviderWithExplicitAndImplicitApis(TestCase):
                 stage_name="Prod")
         ]
 
-        provider = ApiProvider(self.template)
+        provider = SamApiProvider(self.template)
         assertCountEqual(self, expected_apis, provider.apis)
 
     def test_binary_media_types_with_rest_api_id_reference(self):
@@ -934,7 +955,7 @@ class TestSamApiProviderWithExplicitAndImplicitApis(TestCase):
                 stage_name="Prod")
         ]
 
-        provider = ApiProvider(self.template)
+        provider = SamApiProvider(self.template)
         assertCountEqual(self, expected_apis, provider.apis)
 
 
@@ -970,7 +991,7 @@ class TestSamStageValues(TestCase):
                 }
             }
         }
-        provider = ApiProvider(template)
+        provider = SamApiProvider(template)
         api1 = Api(path='/path', method='GET', function_name='NoApiEventFunction', cors=None, binary_media_types=[],
                    stage_name='dev',
                    stage_variables=None)
@@ -1012,7 +1033,7 @@ class TestSamStageValues(TestCase):
                 }
             }
         }
-        provider = ApiProvider(template)
+        provider = SamApiProvider(template)
         api1 = Api(path='/path', method='GET', function_name='NoApiEventFunction', cors=None, binary_media_types=[],
                    stage_name='dev',
                    stage_variables={
@@ -1098,7 +1119,8 @@ class TestSamStageValues(TestCase):
                 }
             }
         }
-        provider = ApiProvider(template)
+
+        provider = SamApiProvider(template)
 
         result = [f for f in provider.get_all()]
